Got Basic Ray Tracing Working!
Basic tracing works and generates an image with something in it which is a milestone in itself. There is no shading, no sampling and no antialiasing. This is just putting the pieces together for future additions. Added a new multi object tracer class and a scene renderer class. Also added a view plane class.